+++ aquaculture-server/src/main/java/com/server/aquacultureserver/service/impl/BaseBreedServiceImpl.java
@@ -5,6 +5,7 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.server.aquacultureserver.domain.BaseBreed;
 import com.server.aquacultureserver.mapper.BaseBreedMapper;
 import com.server.aquacultureserver.service.BaseBreedService;
+import com.server.aquacultureserver.utils.UserContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -51,16 +52,28 @@ public class BaseBreedServiceImpl implements BaseBreedService {
     
     @Override
     public boolean saveBreed(BaseBreed breed) {
+        // 普通操作员不能新增品种
+        if (UserContext.isOperator()) {
+            throw new RuntimeException("普通操作员无权新增品种");
+        }
         return breedMapper.insert(breed) > 0;
     }
     
     @Override
     public boolean updateBreed(BaseBreed breed) {
+        // 普通操作员不能修改品种
+        if (UserContext.isOperator()) {
+            throw new RuntimeException("普通操作员无权修改品种");
+        }
         return breedMapper.updateById(breed) > 0;
     }
     
     @Override
     public boolean deleteBreed(Long breedId) {
+        // 普通操作员不能删除品种
+        if (UserContext.isOperator()) {
+            throw new RuntimeException("普通操作员无权删除品种");
+        }
         return breedMapper.deleteById(breedId) > 0;
     }
 }
